remove useless code, add marlin fingerprint
Remove code that gets (looks what they are) properties that are (probably) not in this phone. Add marlin fingerprint that can help with safetynet. Someone will have to look if I have messed up the code. Also someone will have to look if this actually sets the props. Also if this causes problems I can move these props to the system.prop file.