remove useless code, add marlin fingerprint
remove useless code that gets properties (looks at what they are) and based on that changes what it need (or what someone programmed it to do) this info however isn't in this phone so it's not needed. Also someone will have to look at if I haven't messed something up or if this even makes sense. Setting marlin fingerprint may help pass safetynet